Fix fact that we aren't writing first tags' commits
If a previous tag doesn't exist, we should be checking the project's
first commit. This may be introducing a bug in which we error out if
a new tag doesn't exist. If so we'll have to output some kind of error
to the user that they haven't added a new tag. That will be a future
commit