Fix error when tracking commits
We're splitting the message in accordance
with the formatting. But if the message has a
dash in it - we weren't logging the entire message.

There is actually more of an issue here - that is -
if the user changes the format of the git log in the
configuration. This would not work properly. Will look
into.

Fixes #60