Fix error when tracking commits
We're splitting the message in accordance
with the formatting. But if the message has a
dash in it - we weren't logging the entire message.

There is actually more of an issue here - that is -
if the user changes the format of the git log in the
configuration. This would not work properly. Will look
into.

Fixes #60

This also removes the ability to change the `log_format` since removing
the hyphen from the format could result in unexpected parsing bugs.

Might be worth adding back in the future. For now, let's be content with
setting a standard way of presenting the logs. This is a breaking change with
non backward compatibility, but since no one uses this gem - 
just making a minor version patch.