[Refactor] Delegate all config related methods in one module
This refactors our delegations to an instance of our configuration.
Rather than having it scattered about in each individual file,
this creates a single file with all the delegations we need.

This also prefixes the delegated methods with "config_" in an
effort to make it clear where that method is coming from/referencing